Refactor: the architecture of the application
The folders were first organized by adapters, loaders and viewmodels but after
seeing a discussion on Slack, I've decided to change to component type: MovieDetail, Movies, MovieTrailers and MovieReviews.
The model, database and webservices I left untouched because I felt they were in the right place.
I've tried to refactor MovieActivity and MovieDetail Activity to their respective folders but it was giving a warning were
the application could get a lot of errors since a lot views and other parts were not able to bind properly.